Diagonal estimation
